+++ R/plotmod_helpers_2.R
@@ -69,3 +69,69 @@ x_for_w_categorical <- function(w_i, mf, x, w, row_id = FALSE, ...) {
         return(x1)
       }
   }
+
+#' @noRd
+# Rescale data such that the means and SDs are those required
+
+scale_by_implied <- function(data_original, implied) {
+    data_new <- data_original
+    for (x in colnames(data_original)) {
+        raw_v <- data_original[, x]
+        mean_v <- mean(raw_v)
+        sd_v <- stats::sd(raw_v)
+        sd_v_implied <- sqrt(implied$cov[x, x])
+        if (!(all(is.na(implied$mean)) || is.null(implied$mean))) {
+            mean_v_implied <- implied$mean[x]
+          } else {
+            mean_v_implied <- mean_v
+          }
+        z_v <- scale(data_original[, x])[, 1]
+        new_v <- z_v * sd_v_implied + mean_v_implied
+        data_new[, x] <- new_v
+      }
+    data_new
+  }
+
+#' @noRd
+# Add fill-in latent variable data
+
+add_fillin_lv <- function(data_original, cov_lv, mean_lv = NULL) {
+    lv <- colnames(cov_lv)
+    if (is.null(mean_lv)) {
+        mean_lv <- rep(0, length(lv))
+        names(mean_lv) <- lv
+      }
+    n <- nrow(data_original)
+    dat_lv <- MASS::mvrnorm(n, mu = mean_lv, Sigma = cov_lv, empirical = TRUE)
+    data_new <- cbind(data_original, dat_lv)
+  }
+
+#' @noRd
+# Add implied stats of lv
+
+add_lv_implied <- function(implied_stats, cov_lv, mean_lv = NULL) {
+    lv <- colnames(cov_lv)
+    if (is.null(mean_lv)) {
+        mean_lv <- rep(0, length(lv))
+        names(mean_lv) <- lv
+      }
+    p1 <- nrow(implied_stats$cov)
+    p2 <- length(lv)
+    zero <- matrix(0, nrow = p1, ncol = p2)
+    colnames(zero) <- lv
+    rownames(zero) <- rownames(implied_stats$cov)
+    new_cov <- rbind(cbind(implied_stats$cov, zero),
+                     cbind(t(zero), cov_lv))
+    class(new_cov) <- class(implied_stats$cov)
+    if (!is.null(implied_stats$mean)) {
+        mean_ov <- implied_stats$mean
+      } else {
+        mean_ov <- rep(0, p1)
+        names(mean_ov) <- colnames(implied_stats$cov)
+      }
+    new_mean <- c(mean_ov,
+                  mean_lv)
+    class(new_mean) <- class(mean_lv)
+    out <- list(cov = new_cov,
+                mean = new_mean)
+  }